refactor: inicializando o projeto - personalização de pastas
